+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog for Typopo
 
 
+
+## 2.5.7 //
+
+### 💪 Improvements
+- TBD exceptions on filenames 
+
+### 🐛 Fixes
+- Fix (#73): the part of the file path was falsely identified as an ordinal number in `sk`, `cs`, `rue`, `de-de` locales
+- Fix: Q1 ’23 was falsely changed to Q1’23
+
+### 🔨 Maintenance
+- Write additional tests for exceptions: URLs, filenames, and emails
+- Change packaging from npm to pnpm
+- Update NPM packages to their latest versions
+
+
+
 ## 2.5.6 // 2023-12-23
 
 ### 🐛 Fixes